V.3.1 для iPad устранена ошибка при появлении в Чате окна EULA.
Устранена ошибка:
Thread 1: Exception: "Your application has presented a UIAlertController (<UIAlertController: 0x7fefa8069400>) of style UIAlertControllerStyleActionSheet from UINavigationController (<UINavigationController: 0x7fefa7840000>). The modalPresentationStyle of a UIAlertController with this style is UIModalPresentationPopover. You must provide location information for this popover through the alert controller's popoverPresentationController. You must provide either a sourceView and sourceRect or a barButtonItem.  If this information is not known when you present the alert controller, you may provide it in the UIPopoverPresentationControllerDelegate method -prepareForPopoverPresentation."